Refactor requirements orchestrator into modular flows
